backends/pisi: Override pisi.ui.notify to get callbacks for common ops
We can now emit events from packagekit when we are removing, installing,
downloading or upgrading packages by getting a callback from pisi events.

This allows up to remove the previous hacks.

This commit relies on a change in pisi.fetcher.fetch to emit a `pisi.ui.downloading`
event.

To improve this further we could get a callback from `plan_install_pkg_names`, `plan_upgrade`,
and `plan_remove` in order to set a detailed progress bar for those ops.